Move severity score to vthelper file
