+++ ospd_openvas/daemon.py
@@ -38,7 +38,6 @@ from ospd.ospd import OSPDaemon
 from ospd.scan import ScanProgress
 from ospd.server import BaseServer
 from ospd.main import main as daemon_main
-from ospd.cvss import CVSS
 from ospd.vtfilter import VtsFilter
 from ospd.resultlist import ResultList
 
@@ -969,25 +968,6 @@ class OSPDopenvas(OSPDaemon):
 
         self.sort_host_finished(scan_id, finished_hosts)
 
-    def get_severity_score(self, vt_aux: dict) -> Optional[float]:
-        """Return the severity score for the given oid.
-        Arguments:
-            vt_aux: VT element from which to get the severity vector
-        Returns:
-            The calculated cvss base value. None if there is no severity
-            vector or severity type is not cvss base version 2.
-        """
-        if vt_aux:
-            severity_type = vt_aux['severities'].get('severity_type')
-            severity_vector = vt_aux['severities'].get('severity_base_vector')
-
-            if severity_type == "cvss_base_v2" and severity_vector:
-                return CVSS.cvss_base_v2_value(severity_vector)
-            elif severity_type == "cvss_base_v3" and severity_vector:
-                return CVSS.cvss_base_v3_value(severity_vector)
-
-        return None
-
     def report_openvas_results(self, db: BaseDB, scan_id: str) -> bool:
         """Get all result entries from redis kb."""
 
@@ -1086,7 +1066,7 @@ class OSPDopenvas(OSPDaemon):
                 )
 
             elif msg[0] == 'ALARM':
-                rseverity = self.get_severity_score(vt_aux)
+                rseverity = vthelper.get_severity_score(vt_aux)
                 res_list.add_scan_alarm_to_list(
                     host=current_host,
                     hostname=rhostname,
